requirements.txt
aiohttp==3.7.4.post0
appdirs==1.4.4
async-timeout==3.0.1
attrs==21.2.0
chardet==4.0.0
clikit==0.6.2
crashtest==0.3.1
cycler==0.10.0
GetDist==1.3.2
httpstan==4.6.0
idna==3.2
kiwisolver==1.3.2
lz4==3.1.3
marshmallow==3.13.0
matplotlib==3.4.3
multidict==5.2.0
numpy==1.21.2
pastel==0.2.1
Pillow==8.3.2
pylev==1.4.0
pyparsing==2.4.7
pysimdjson==3.2.0
pystan==3.3.0
python-dateutil==2.8.2
PyYAML==5.4.1
scipy==1.7.1
six==1.16.0
typing-extensions==3.10.0.2
webargs==8.0.1
yarl==1.7.0
